Remove Post-Meritocracy Manifesto Reason: Can't make up my mind re:humanetech, plus maintainer follows questionable practices.
After I PR'ed a Dutch translation I noticed the toxicity in
the issue tracker and started a long discussion with one of
the more reasonable objectors. We found many improvements
and things to consider + included a mention to the maintainer.

No response.

Only later I found that the entire discussion was silently
erased from the issue tracker, without giving any
notification. This is bad practice and reason for removal.